requirements.txt
langchain==0.0.340
openai==1.3.5
python-dotenv==1.0.0
pytube==15.0.0
streamlit==1.28.2
tiktoken==0.5.1
youtube-transcript-api==0.6.1

